Add support for unknown commands and numerics
This adds the UnknownCommand class, which proxies to a Command class
if a command of the same name is later declared; otherwise, it looks
like a basic Command instance with no declared arguments.  This also
adds support for numerics, and an implementation of get_command()
that returns an UnknownCommand when necessary.